update logic and tests

+++ src/Projects/Projects.Application/Features/DeleteProject/DeleteProjectCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using LanguageExt.Common;
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Projects.Application.LogicValidation;
 using Projects.Infrastructure.Interfaces;
 
 namespace Projects.Application.Features.DeleteProject
@@ -8,12 +9,14 @@ namespace Projects.Application.Features.DeleteProject
     public class DeleteProjectCommandHandler : IRequestHandler<DeleteProjectCommand, Result<bool>>
     {
         private readonly ILogger<DeleteProjectCommandHandler> _logger;
+        private readonly IBusinessLogicValidator _businessLogicValidator;
         private readonly IUnitOfWork _unitOfWork;
 
-        public DeleteProjectCommandHandler(ILogger<DeleteProjectCommandHandler> logger, IUnitOfWork unitOfWork)
+        public DeleteProjectCommandHandler(ILogger<DeleteProjectCommandHandler> logger, IUnitOfWork unitOfWork, IBusinessLogicValidator businessLogicValidator)
         {
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
             _unitOfWork = unitOfWork ?? throw new ArgumentNullException(nameof(unitOfWork));
+            _businessLogicValidator = businessLogicValidator;
         }
 
         public async Task<Result<bool>> Handle(DeleteProjectCommand request, CancellationToken cancellationToken)
@@ -25,6 +28,17 @@ namespace Projects.Application.Features.DeleteProject
 
             try
             {
+                var deleteProjectValidationRequest = await _businessLogicValidator.ValidateAsync(
+                    new DeleteProjectValidationRequest(request.ProjectId, request.TenantId), cancellationToken)
+                    .ConfigureAwait(false);
+                if (deleteProjectValidationRequest != null)
+                {
+                    _logger.LogInformation("Project {Project} under account {Account} cannot be deleted due to business logic rules violation: {Response}",
+                                                                  request.ProjectId, request.TenantId, deleteProjectValidationRequest);
+
+                    return new Result<bool>(deleteProjectValidationRequest);
+                }
+
                 await _unitOfWork.TaskItemsRepository.DeleteAllAsync(request.ProjectId, request.TenantId, cancellationToken);
 
                 await _unitOfWork.ProjectsRepository.DeleteProjectAsync(request.ProjectId, request.TenantId, cancellationToken);
